Fixed ej1 bug

+++ TP4/src/distance.py
@@ -2,6 +2,6 @@ import numpy as np
 
 
 def euclidean(a, b):
-    a = np.array(a)
-    b = np.array(b)
+    a = np.array(a).astype(float)
+    b = np.array(b).astype(float)
     return np.linalg.norm(a - b)
